Add files via upload
ReproBiCGSTAB_ MPIori. c is to call MPI_ Allreduce's parallel preprocessing BiCGSTAB;
ReproBiCGSTAB_ MPImpfr. C is a serial high-precision preprocessing BiCGSTAB based on MPFR, which serves as the baseline.
ReproBiCGSTAB_ MPIexblas. C is the parallel pre-processing BiCGSTAB based on ExBLAS, and the distributed dot product is the dot product called ExBLAS.
ReproBiCGSTAB_ MPIreprobla. C is the parallel preprocessing BiCGSTAB based on ReproBLAS, and the distributed dot product is the dot product called ReproBLAS.